REF: Factor out ServicePatcher in scripts
