+++ scripts/process_questions_across_versions/process_question.py
@@ -3,18 +3,19 @@ import json
 import os
 import sys
 import tempfile
-from unittest.mock import patch
 
 from octue.resources import Datafile, Dataset, Manifest
 
+from ..utils import ServicePatcher
+
 
 try:
-    from octue.cloud.emulators._pub_sub import MESSAGES, MockService, MockSubscriber, MockSubscription, MockTopic
+    from octue.cloud.emulators._pub_sub import MESSAGES, MockService
 except ModuleNotFoundError:
     try:
-        from tests.cloud.pub_sub.mocks import MESSAGES, MockService, MockSubscriber, MockSubscription, MockTopic
+        from tests.cloud.pub_sub.mocks import MESSAGES, MockService
     except ModuleNotFoundError:
-        from octue.cloud.emulators.pub_sub import MockService, MockSubscriber, MockSubscription, MockTopic, MESSAGES
+        from octue.cloud.emulators.pub_sub import MockService, MESSAGES
 
 from octue.resources.service_backends import GCPPubSubBackend
 
@@ -32,30 +33,6 @@ output_manifest = Manifest(
         )
     }
 )
-
-
-class ServicePatcher:
-    def __init__(self, patches=None):
-        self.patches = patches or [
-            patch("octue.cloud.pub_sub.service.Topic", new=MockTopic),
-            patch("octue.cloud.pub_sub.service.Subscription", new=MockSubscription),
-            patch("google.cloud.pubsub_v1.SubscriberClient", new=MockSubscriber),
-        ]
-
-    def __enter__(self):
-        """Start the patches and return the mocks they produce.
-
-        :return list(unittest.mock.MagicMock):
-        """
-        return [patch.start() for patch in self.patches]
-
-    def __exit__(self, *args, **kwargs):
-        """Stop the patches.
-
-        :return None:
-        """
-        for p in self.patches:
-            p.stop()
 
 
 class MockAnalysis:
