+++ scripts/record_questions_across_versions/record_question.py
@@ -8,14 +8,16 @@ import pkg_resources
 from octue.resources import Datafile, Dataset, Manifest
 from octue.utils.encoders import OctueJSONEncoder
 
+from ..utils import ServicePatcher
+
 
 try:
-    from octue.cloud.emulators._pub_sub import MockService, MockSubscriber, MockSubscription, MockTopic
+    from octue.cloud.emulators._pub_sub import MockService
 except ModuleNotFoundError:
     try:
-        from tests.cloud.pub_sub.mocks import MockService, MockSubscriber, MockSubscription, MockTopic
+        from tests.cloud.pub_sub.mocks import MockService
     except ModuleNotFoundError:
-        from octue.cloud.emulators.pub_sub import MockService, MockSubscriber, MockSubscription, MockTopic
+        from octue.cloud.emulators.pub_sub import MockService
 
 from octue.resources.service_backends import GCPPubSubBackend
 
@@ -33,30 +35,6 @@ input_manifest = Manifest(
         )
     }
 )
-
-
-class ServicePatcher:
-    def __init__(self, patches=None):
-        self.patches = patches or [
-            patch("octue.cloud.pub_sub.service.Topic", new=MockTopic),
-            patch("octue.cloud.pub_sub.service.Subscription", new=MockSubscription),
-            patch("google.cloud.pubsub_v1.SubscriberClient", new=MockSubscriber),
-        ]
-
-    def __enter__(self):
-        """Start the patches and return the mocks they produce.
-
-        :return list(unittest.mock.MagicMock):
-        """
-        return [patch.start() for patch in self.patches]
-
-    def __exit__(self, *args, **kwargs):
-        """Stop the patches.
-
-        :return None:
-        """
-        for p in self.patches:
-            p.stop()
 
 
 class QuestionRecorder:
